% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibrated_stic_data}
\alias{calibrated_stic_data}
\title{Example calibrated STIC output data.}
\format{
## `calibrated_stic_data`
A data frame with 1000 rows and 4 columns:
\describe{
  \item{datetime}{Date and time of measurement.}
  \item{condUncal}{Raw uncalibrated conductivity recorded by STIC logger.}
  \item{tempC}{Temperature recorded by STIC logger.}
  \item{SpC}{Specific conductance calculated using `apply_calibration` function.}
}
}
\source{
AIMS project data.
}
\usage{
calibrated_stic_data
}
\description{
Calibrated STIC data used for function examples.
}
\keyword{datasets}
