% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_standard_data}
\alias{calibration_standard_data}
\title{Example calibration STIC lab data.}
\format{
## `calibration_standard_data`
A data frame with 4 rows and 3 columns:
\describe{
  \item{sensor}{Serial number for STIC sensor.}
  \item{standard}{Specific conductance (`SpC`) standard values used for soaking STIC.}
  \item{condUncal}{Uncalibrated conductivity recorded by STIC when soaked in each standard.}
}
}
\source{
AIMS project data.
}
\usage{
calibration_standard_data
}
\description{
Example calibration data for STIC sensor for conversion from uncalibrated conductivity to specific conductivity (`SpC`).
}
\keyword{datasets}
