% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classified_df}
\alias{classified_df}
\title{Example classified STIC output data.}
\format{
## `classified_df`
A data frame with 1000 rows and 5 columns:
\describe{
  \item{datetime}{Date and time of measurement.}
  \item{condUncal}{Raw uncalibrated conductivity recorded by STIC logger.}
  \item{tempC}{Temperature recorded by STIC logger.}
  \item{SpC}{Specific conductance calculated using `apply_calibration` function.}
  \item{wetdry}{Classified STIC data created by `classify_wetdry` function.}
}
}
\source{
AIMS project data.
}
\usage{
classified_df
}
\description{
Classified STIC data used for function examples.
}
\keyword{datasets}
