% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field_obs}
\alias{field_obs}
\title{Example field observations that could be compared to classified STIC data.}
\format{
## `field_obs`
A data frame with 5 rows and 3 columns:
\describe{
  \item{datetime}{Date and time of field observation.}
  \item{wetdry}{Field observation of stream water status (`wet` or `dry`).}
  \item{SpC}{Field observations of specific conductance.}
}
}
\source{
Made up data.
}
\usage{
field_obs
}
\description{
Example field observations that could be compared to classified STIC data.
}
\keyword{datasets}
