% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_STIC.R
\name{load_STIC}
\alias{load_STIC}
\title{Import raw STIC sensor data}
\usage{
load_STIC(data = NULL, dir = NULL)
}
\arguments{
\item{data}{A dataframe}

\item{dir}{A directory with raw STIC files}
}
\value{
the formatted data
}
\description{
load_STIC() is a tool to import and tidy raw STIC sensor data.
}
\examples{
clean_STIC <- load_STIC(data = "https://raw.githubusercontent.com/matthewberens/STICr/data/raw_hobo_export.csv")


}
