% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIC_imports.R
\name{loadSTIC}
\alias{loadSTIC}
\alias{loadSTICdir}
\title{Import raw STIC sensor data}
\usage{
loadSTIC(rawSTIC)

loadSTICdir(dirSTIC)
}
\arguments{
\item{rawSTIC}{A dataframe}

\item{dirSTIC}{A directory with raw STIC files}
}
\value{
the formatted data
}
\description{
loadSTIC() is a tool to import and tidy raw STIC sensor data.
`loadSTICdir()`imports and tidies a directory containing multiple raw STIC sensor files.
}
