% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_STIC.R
\name{plot_STIC}
\alias{plot_STIC}
\title{Plotting STIC Sensor Data}
\usage{
plot_STIC(
  data,
  site = NULL,
  date_breaks = "2 weeks",
  date_labels = "\%b-\%d",
  xlab = NULL,
  plotly = FALSE
)
}
\arguments{
\item{data}{A dataframe}

\item{site}{Location}

\item{date_labels}{}
}
\value{
the formatted data
}
\description{
`plot_STIC` is a function to make a ggplot plot with STIC data.
}
