% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_stic_data}
\alias{tidy_stic_data}
\title{Example tidied STIC output data.}
\format{
## `tidy_stic_data`
A data frame with 1000 rows and 3 columns:
\describe{
  \item{datetime}{Date and time of measurement.}
  \item{condUncal}{Raw uncalibrated conductivity recorded by STIC logger.}
  \item{tempC}{Temperature recorded by STIC logger.}
}
}
\source{
AIMS project data.
}
\usage{
tidy_stic_data
}
\description{
Example tidied STIC data for input to calibration and classification process.
}
\keyword{datasets}
