% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_calibration.R
\name{apply_calibration}
\alias{apply_calibration}
\title{apply_calibration}
\usage{
apply_calibration(stic_data, calibration)
}
\arguments{
\item{stic_data}{A data frame with a column named `conductivity_uncal`, for example as produced by the function `tidy_hobo_data`.}

\item{calibration}{a model object relating `conductivity_uncal` to a standard of some sort, for example as produced by the function `get_calibration`.}
}
\value{
The same data frame as input, except with a new column called `SpC`. This will be in the same units as the data used to develop the model calibration.
}
\description{
apply_calibration
}
\examples{
stic_data <- read.csv("data/TestData_20946471_Clean.csv")
calibration_data <- read.csv("data/TestData_20946471_Calibration.csv")
calibration <- get_calibration(calibration_data, method = "linear")
calibrated_df <- apply_calibration(stic_data, calibration)
head(calibrated_df)

}
