% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_wetdry.R
\name{classify_wetdry}
\alias{classify_wetdry}
\title{classify_wetdry}
\usage{
classify_wetdry(stic_data, classify_var = "spc", threshold = 200)
}
\arguments{
\item{stic_data}{A data frame with STIC data, such as that produced by `apply_calibration` or `tidy_hobo_data`}

\item{classify_var}{A column name of the column in data frame you want to use for classification.}

\item{threshold}{An absolute numerical threshold for classifying wet vs dry}
}
\value{
The same data frame as input, but with a new column called `wetdry`.
}
\description{
This is a function to classify STIC data into a binary "wet" and "dry" column. Data can be classified according to any classification variable defined by the user. User can choose one of two methods for classification: either an absolute numerical threshold or as a chosen percentage of the maximum value of the classification variable.
}
\examples{
stic_data <- calibrated_stic_data
classified_df <- classify_wetdry(stic_data, classify_var = "spc", threshold = 200)
head(classified_df)
}
