% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_wetdry.R
\name{classify_wetdry}
\alias{classify_wetdry}
\title{classify_wetdry}
\usage{
classify_wetdry(stic_data, classify_var = "spc", threshold = 200)
}
\arguments{
\item{stic_data}{A data frame with STIC data, such as that produced by `apply_calibration` or `tidy_hobo_data`}

\item{classify_var}{A column name of the column in data frame you want to use for classification.}

\item{threshold}{An absolute numerical threshold for classifying wet vs dry}
}
\value{
The same data frame as input, but with a new column called wetdry
}
\description{
classify_wetdry
}
