% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_wetdry.R
\name{classify_wetdry}
\alias{classify_wetdry}
\title{classify_wetdry}
\usage{
classify_wetdry(stic_data, classify_var = "spc", threshold = 200)
}
\arguments{
\item{stic_data}{A data frame with STIC data, such as that produced by `apply_calibration` or `tidy_hobo_data`}

\item{classify_var}{A column name of the column in data frame you want to use for classification.}

\item{threshold}{This is the user-defined threshold for determining wet versus dry based on the designated classification variable. If using the "absolute" method, the threshold will be a value in the same units as the designated classification variable. If using the "percent" method, the value will be a decimal percentage of the max value of the classification variable in the data frame. Values above this proportion of the maximum will be designated as wet.}

\item{method}{User chooses which classification method used to generate the binary data. "absolute" uses an absolute numerical threshold for classifying wet vs dry. "percent" uses a threshold based on a given percentage of the maximum value of the classification variable in the data frame.}
}
\value{
The same data frame as input, but with a new column called `wetdry`.
}
\description{
This is a function to classify STIC data into a binary "wet" and "dry" column. Data can be classified according to any classification variable defined by the user. User can choose one of two methods for classification: either an absolute numerical threshold or as a chosen percentage of the maximum value of the classification variable.
}
\examples{
stic_data <- calibrated_stic_data
classified_df <- classify_wetdry(stic_data, classify_var = "spc", method = "absolute", threshold = 200)
head(classified_df)
}
