% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calibration.R
\name{get_calibration}
\alias{get_calibration}
\title{get_calibration}
\usage{
get_calibration(calibration_data, method = "linear")
}
\arguments{
\item{calibration_data}{STIC calibration data frame with columns "standard" and "conductivity_uncal"}

\item{method}{method for creating interpolation: "linear" (default) or "exponential"}
}
\value{
fitted model object
}
\description{
get_calibration
}
\examples{
calibration_data <- read.csv("data/TestData_20946471_Calibration.csv")
lm_calibration <- get_calibration(calibration_data, method = "linear")
View(lm_calibration)

}
