% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calibration.R
\name{get_calibration}
\alias{get_calibration}
\title{get_calibration}
\usage{
get_calibration(calibration_data, method = "linear")
}
\arguments{
\item{calibration_data}{STIC calibration data frame with columns "standard" and "conductivity_uncal"}

\item{method}{method for creating interpolation: "linear" (default) or "exponential"}
}
\value{
A fitted model object relating SpC to the uncalibrated conductivity values measured by the STIC
}
\description{
Function to fit SpC standards and uncalibrated conductivity measured by the STIC to a model object. This model will then be used to predict SpC values using 'apply_calibration'
}
\examples{
calibration_data <- calibration_standard_data
lm_calibration <- get_calibration(calibration_data, method = "linear")
View(lm_calibration)

}
