% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_stic_data.R
\name{qaqc_stic_data}
\alias{qaqc_stic_data}
\title{qaqc_stic_data}
\usage{
qaqc_stic_data(
  stic_data,
  spc_neg_correction = TRUE,
  inspect_classification = TRUE,
  anomaly_size = 4,
  window_size = 1000,
  concatenate_flags = TRUE
)
}
\arguments{
\item{stic_data}{A data frame with classified STIC data, such as that produced by `classify_wetdry`}

\item{spc_neg_correction}{a logical argument indicating whether the user would like to correct negative SPC values resulting from the calibration process to 0.
The character code associated with this correction is "N".}

\item{inspect_classification}{a logical argument indicating whether the user would like to identify instances in which either a wet or dry reading is surrounded on both sides by 1000 or more observations of its opposite.
This operation is meant to identify potentially suspect binary wet/dry data points for further examination.
The character code associated with this operation is "A".}

\item{anomaly_size}{a numeric argument specifying the maximum size (i.e., number of observations) of a clustered group of points that can be flagged as an anomaly}

\item{window_size}{a numeric argument specifying the minimum size (i.e., number of observations) that the anomaly must be surrounded by in order to be flagged}

\item{concatenate_flags}{a logical argument indicating whether the user would like to combine the character codes generated into a single QAQC flag column.}
}
\value{
The same data frame as input, but with new QAQC columns or a single, concatenated QAQC column
}
\description{
This function provides multiple options for QAQC flagging of processed and classified STIC data frames, such as those generated by the `classify_wetdry function`.
Users can select which operations are to be performed, and a character flag is added as an addition column in the input data frame.
Users also have to option to concatenate multiple flags into one column. QAQC options currently include...
}
\examples{
qaqc_df <- qaqc_stic_data(classified_df, spc_neg_correction = TRUE, inspect_classification = TRUE, concatenate_flags = TRUE)
}
