% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_hobo_data.R
\name{tidy_hobo_data}
\alias{tidy_hobo_data}
\title{tidy_hobo_data}
\usage{
tidy_hobo_data(infile, outfile = FALSE)
}
\arguments{
\item{infile}{filename (including path if needed) for the raw HOBO data on your computer.}

\item{outfile}{filename (including path if needed) to save the tidied. Defaults to FALSE, in which case tidied data will not be saved.}
}
\value{
a tidied data frame
}
\description{
tidy_hobo_data
}
\examples{
tidy_df <- tidy_hobo_data("data/TestData_20946471_Raw.csv")
head(tidy_df)

}
