% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_stic_data.R
\name{validate_stic_data}
\alias{validate_stic_data}
\title{validate_stic_data.R}
\usage{
validate_stic_data(
  stic_data,
  include_SpC = TRUE,
  include_classification = TRUE
)
}
\arguments{
\item{stic_data}{classified STIC data frame with the variable names of that produced by `classify_wetdry`, with the addition of two additional columns called `field_SpC` and `field_classification`}

\item{include_SpC}{logical argument where the user specifies if they want a matrix for SpC measurements}

\item{include_classification}{logical argument where the user specifies if they want a matrix for SpC measurements}
}
\value{
one or two confusion matrices
}
\description{
This function creates two confusion matrices from a classified STIC data frame with the variable names of that produced by `classify_wetdry`, with the addition of two additional columns called `field_SpC` and `field_classification`
}
\examples{
data_validation_confusion_matrix <- validate_stic_data(stic_data = classified_stic_with_field_obs, include_SpC = TRUE, include_classification = TRUE)
}
